Extract add trip form to own component
