+++ frontends/vue-webpack/src/components/addtripform.vue
@@ -1,0 +1,61 @@
+<template>
+  <div id="addtripform" style="display: block;">
+    <form name="trip-add-form" id="trip-add-form">
+      <div class="form-group row">
+        <label for="tripDate" class="col-sm-2 col-form-label">Date</label>
+        <div class="col-sm-10">
+          <datepicker :disabled="disabled" :value="date()" inline bootstrapStyling monday-first required></datepicker>
+        </div>
+      </div>
+      <div class="form-group row">
+        <label class="col-sm-2 col-form-label">Distance</label>
+        <div class="col-sm-10">
+          <input type="text" class="form-control" id="distanceKm" name="distanceKm" value="">
+        </div>
+      </div>
+      <div class="form-group row">
+        <div class="offset-sm-2 col-sm-10">
+          <input type="button" class="btn btn-danger" value="Cancel">
+          <input type="button" @click="submitTrip()" class="btn btn-primary" value="Submit"></button>
+        </div>
+      </div>
+    </form>
+    <button class="btn btn-primary float-sm-right" style="margin-bottom: 10px;">
+      <icon name="plus" aria-hidden="true"></icon> Add a new trip
+    </button>
+  </div>
+</template>
+
+<script>
+import Datepicker from 'vuejs-datepicker'
+
+export default {
+  name: 'addtripform',
+  data() {
+    return {
+      disabled: {
+        to: new Date(2017, 4, 1), // Disable all dates up to specific date
+        from: new Date(2017, 4, 31), // Disable all dates after specific date
+        days: [6, 0] // Disable Saturday's and Sunday's
+      },
+      date: function() {
+        var now = new Date()
+        var d = new Date(now.getFullYear(), 4, now.getDate())
+        return d
+      }
+    }
+  },
+  components: {
+    Datepicker
+  },
+  methods: {
+    submitTrip: function() {
+      console.log('submitting trip')
+    }
+  }
+}
+</script>
+
+<style>
+
+</style>
