+++ frontends/vue-webpack/src/components/teaminvitations.vue
@@ -1,6 +1,6 @@
 <template>
   <b-card header="Team invitations">
-      <p>You currently have no team invitations</p>
+    <p>You currently have no team invitations</p>
   </b-card>
 </template>
